서버 업데이트(ver.1.0.1)
서버 데이터 수신 기능 구현

+++ server/BPS_server/BPS_server/Server.h
@@ -16,8 +16,8 @@ struct ClientId {
 
 //객체의 2차원 임시좌표
 struct Point2D {
-	float position_x;
-	float position_y;
+	int position_x;
+	int position_y;
 };
 
 //객체의 2차원 가속도(임시)
